Try to start ffado-dbus-server if it can't be found on the DBus.
This is probably not the right solution, it would be better to register the server as a dbus-service and request that service so dbus autostarts it.
And the error handling isn't that nice either. if ffado-dbus-server can't be started, it keeps on asking until you press "No".
And probably the time to wait for the next try is to short or to long...


git-svn-id: http://subversion.ffado.org/ffado@804 2be59082-3212-0410-8809-b0798e1608f0