Mix it baby!
A first matrix with dials (extra for ppalmers:-P). For screen-space its limited to 8x8 elements currently (or less, if the device has less).
Have to think about how to hide channels that aren't used. Or hide groups of channels with similar names. Hint: this would currently result in one group named "FIXME" :-)



git-svn-id: http://subversion.ffado.org/ffado@1636 2be59082-3212-0410-8809-b0798e1608f0