DICE-EAP Mute rather than remove route when disabling a destination
By Philippe Carriere:

Really removing a route is not fully supported by all devices; then
muting is preferable when one just want to set a different connection
state (as part of a modification of one source for another one for a
given destination).

A getMuteId() function is introduced so as to be easily customized for
specific devices, if necessary.

git-svn-id: http://subversion.ffado.org/ffado@2096 2be59082-3212-0410-8809-b0798e1608f0