A series of fixes provided by Andreas Hehn.  These were mostly discovered when attempting to compile FFADO with clang: * libavc: implement a better way to quieten an "unused parameter" warning * bebob: fix a typo when assigning to cachePath * dice: correctly create routes for the ARM audio ports * fireworks-downloader: really zero the "args" variable * saffire-pro: fix a typo when applying sync lock mask
I'm not sure what the correct approach should be for the DICE fix.  Andreas's patch originally commented this out.  The DICE chip does provide 8 ARM audio ports on the router's input and output sides though, so perhaps they ought to be included (although the present entries only seem to cover the router outputs to the ARM, not the returns from it).  Those with DICE-based interfaces are encouraged to test this patch to verify there are no regressions caused by this change.



git-svn-id: http://subversion.ffado.org/ffado@2413 2be59082-3212-0410-8809-b0798e1608f0