[Fireworks] add a workaround for Audiofire12 (and maybe Audiofire8)
EfcGetClockCmd over FCP is often failed or return invalid data. This commit add retry for doEfcOverFCP().

When the trials are completely failed, for sampling rate operation, a fallback to 'input/output plug signal format' command is executed. This command seems to be supported by all models.

But the driver can't know current clock source via 'plug signal format' command. This commit add 'm_current_clock' member to keep the id. This member is updated by setClock(). When getClock() is faled even if this member is not set, fallback to internal clock.


git-svn-id: http://subversion.ffado.org/ffado@2452 2be59082-3212-0410-8809-b0798e1608f0