Clean up: Remove E_Prepared state from IsoHandlerManager.
E_Prepared is only set in one place, IsoHandlerManager::stopHandlers().
This is only called on shutdown, and the IsoHandlerManager is destroyed
shortly after, hence E_Prepared will never actually be tested for.

In addition to being unnecessary, it is also confusingly named, if
anything it should be called E_ShuttingDown, due to where it's set. But
since it is not necessary we remove it instead.

Patch from Kristian Amlie.


git-svn-id: http://subversion.ffado.org/ffado@2591 2be59082-3212-0410-8809-b0798e1608f0