Add the "-b" / "--bypassdbus" command line option to ffado-mixer.  If given, no communication with ffado-dbus-server will be attempted.  Instead, an instance of each supported mixer module will be created with stubs providing placeholder values in place of values obtained from hardware.  This permits rudimentary debugging of mixer modules for interfaces which the developer does not have on hand.
This commit amounts to a minor rework of a patch originally submitted by
Xavier Forestier as part of a large series which is being slowly merged:

  https://sourceforge.net/p/ffado/mailman/message/35457569/.

The most significant difference is that the bypassdbus functionality in
Xavier's patch was a compile option whereas the commit implements it as a
runtime option.  It is felt that this will make things easier for developers
in the long run: mixer modules can be tested without having to reconfigure
ffado from source.  Minor bug fixes were applied to the patch as it was
reworked.

Thanks to Xavier Forestier for the original patch.



git-svn-id: http://subversion.ffado.org/ffado@2670 2be59082-3212-0410-8809-b0798e1608f0