A first pass at optionally supporting PyQt5 in the build system.
This is an initial implementation of changes in the build system to support
the use of PyQt5 when building ffado-mixer.  They are based on the patch
set submitted to ffado-devel by Xavier Forestier in November 2016.  However,
rather than removing PyQt4 support, the idea is that either PyQt4 or PyQt5
can be used.

This patch has been tested on a machine with only PyQt4 installed.  Testing
with only PyQt5 is yet to be done (the author presently does not have access
to such a system).

It is not known how well this patch will work on a machine with both PyQt4
and PyQt5 installed.  r2694 assumes that PyQt4 will be used in this
situation on the basis that it has had the most testing, but it's unclear
whether the build system will select pyuic4 over pyuic5 in this case. 
Further testing will be required, and adjustments will be made if problems
are found.



git-svn-id: http://subversion.ffado.org/ffado@2698 2be59082-3212-0410-8809-b0798e1608f0