matrixmixer.py: attempt to deal with UTF8 string changes in python3.
PyQt5 under python3 does not implement QString and instead relies on the
native python UTF string type.  Attempt to deal with this when drawing mixer
elements which make use of UTF characters.  This has not yet been tested under
python3 due to the author not having access to a system with python3, so
testing may yet prove that the approach taken in this commit is
insufficient.

This commit is based on the suggestion from Xavier Forestier in his patch
set from November 2016 submitted via the ffado-devel mailing list.



git-svn-id: http://subversion.ffado.org/ffado@2699 2be59082-3212-0410-8809-b0798e1608f0