rme: correctly save playback and output mixer settings.
A cut-and-paste oversight at the time the save/restore code was written
resulted in the input matrix values being saved for the playback and output
matrix mixers as well.  Currect this so the right matrix mixer will be read
when writing the respective parts of the saved configuration.

Thanks to Vince who reported the symptoms on ffado-devel and identified the
cause.


git-svn-id: http://subversion.ffado.org/ffado@2709 2be59082-3212-0410-8809-b0798e1608f0