Support the "ppc64le" architecture.
The ppc64le is basically the ppc64 architecture operating in little
endian mode (to port x86 applications easier).

The Fedora build servers are "power8" systems, requiring an extension of the
PPC 64-bit test in order to compile with the correct word size.

Thanks to Orcan Ogetbil for the patch (via ffado-devel mailing list).


git-svn-id: http://subversion.ffado.org/ffado@2717 2be59082-3212-0410-8809-b0798e1608f0