SCons: fix detection of dbus on PyQt5 systems.
The detection of the dbus python modules was qualified on the presence of
pyuic4.  On systems with only PyQt5 this would obviously fail.  Patch from
Orcan Ogetbil via the ffado-devel mailing list.

r2724 inadvertently included a change to SConstruct to import sys.  This is
needed since sys.stdout is referenced in some situations (in particular, if
jack is not installed in the build environment).  This patch was also from
Orcan Ogetbil, via the ffado-devel mailing list.


git-svn-id: http://subversion.ffado.org/ffado@2725 2be59082-3212-0410-8809-b0798e1608f0