Further Python3/PyQt5 compatibility fixes, enhance "About" dialog box.
This patch was prepared by Orcan Ogetbil on Fedora and tested on that
distribution with PyQt4-4.12.1 and PyQt5-5.9.1.  It is primarily concerned
with additional compatibility between Python2/3 and PyQt4/5.  It has also
been tested under python 2 with pyqt 4.11.4.  Further testing under
different python/pyqt environments would be welcomed.

An enhancement to the "About" dialog is also included which adds version
strings and updates the copyright year range.

The patch was submitted to the ffado-devel mailing list.


git-svn-id: http://subversion.ffado.org/ffado@2726 2be59082-3212-0410-8809-b0798e1608f0