Add build-time control over documentation generation.
Up to this point, FFADO documentation has always been generated when 
doxygen is on the system.  Furthermore, the documentation included internal
API details which are not of any interest to users of FFADO.

This patch from Orcan Ogetbil (via ffado-devel) adds the BUILD_DOC scons
parameter which takes one of three values: "none" (don't build any
documentation), "user" (build documentation of interest to users of FFADO)
and "all" (build all documentation).  The default is "none", mostly because
Orcan's distribution (Fedora) has traditionally patched
doc/reference.doxygen.in to exclude documentation.  This is fine since the
documentation is not of interest to a vast majority of people.


git-svn-id: http://subversion.ffado.org/ffado@2733 2be59082-3212-0410-8809-b0798e1608f0