Add comment explaining the "__future__" print_function imports.
While Python 2.7 natively supports the print() function syntax of Python3,
it is convenient to also make use of the "print(..., end='')" feature.  This
is not available under Python2 by default, but can be accessed in Python2
by way of the "from __future__ import print_function" statement.  Include
a brief comment to this effect in the relevant source files to make the
purpose of the "__future__" import clear in future when the need for such
lines might be deemed unnecessary.


git-svn-id: http://subversion.ffado.org/ffado@2750 2be59082-3212-0410-8809-b0798e1608f0