Add support for a ffado-mixer.appdata.xml file.
Distributions are starting to rely on appdata files to present applications
in package management systems.  Add an appdata file for ffado-mixer since
this is the only significant component of FFADO which counts as a user
application.  Thanks to Orcan Ogetbil for raising the issue on ffado-devel
and providing an initial appdata file.

For the moment this appdata file contains no screenshot, mostly because
the screen content is dependent on the firewire interface in use at the
time.  We may revisit this in future if there's a call to do so.


git-svn-id: http://subversion.ffado.org/ffado@2756 2be59082-3212-0410-8809-b0798e1608f0