ffado-mixer.appdata.xml: add a screenshot.
According to the appdata specification at least one screenshot must be
supplied.  This patch adds a URL to such a screenshot.  A single screenshot
will do for now: since the appearance varies significantly depending on the
audio interface in use there isn't a lot of point in providing too many
detailed screenshots.


git-svn-id: http://subversion.ffado.org/ffado@2758 2be59082-3212-0410-8809-b0798e1608f0