ffado-mixer.appdata.xml: add "OR GPLv3" to project licence.
Jano Svitok has determined that the project licence field of the appdata
file can specify multiple licences separated with "OR".  I haven't yet seen
that in the appdata specification documentation but apparently it passes the
validator.  Since many files in FFADO are currently GPLv2 or GPLv3 (without
the "or later" clause) it is fair to add "OR GPL-3.0" to the project licence
field for the moment.

Following the next release, the licence comments in some FFADO files should
be cleaned up to ensure consistency across the project.

Thanks to Jano for the patch (via ffado-devel).


git-svn-id: http://subversion.ffado.org/ffado@2760 2be59082-3212-0410-8809-b0798e1608f0