Fix handling of errno in test-mixer and test-pan.
To maintain compatibility with gcc6, errno can no longer be initialised like
a variable.  To ensure it's always defined, inclusion of cerrno is needed.

These issues were solved in test-volume.cpp and test-enhanced-mixer.cpp
in r2605 (errno initialisation) and r2607 (inclusion of cerrno).  Obviously
the same changes are needed in test-mixer and test-pan., and it's not clear
how this was missed in the earlier patches.  It is probably related to the
fact that these test programs are rarely compiled these days.

Patch from Dave Plater.


git-svn-id: http://subversion.ffado.org/ffado@2766 2be59082-3212-0410-8809-b0798e1608f0