ffado-mixer: don't try loading PyQt4 if Python3 is in use.
While defaulting to PyQt4 is reasonable if Python2 is in use, this is not
generally applicable under Python3.  In the Python3 case, PyQt5 should be
used because the python3 code assumes the newer API offered by PyQt5.  This
patch applies this logic in a way which works on Python3 systems which have
both PyQt4 and PyQt5 installed.

Thanks to David Runge for notifying us about this issue and for an initial
patch.


git-svn-id: http://subversion.ffado.org/ffado@2774 2be59082-3212-0410-8809-b0798e1608f0