mixer-qt4: fix for building with scons under python 3.
This is a quick fix to use os.walk instead of os.path.walk, since the latter
is gone in Python 3.  Python 2 has both, so it should work fine.

Note that this is about running SCons itself on Python 3, not running
ffado-mixer et al on Python 3.

There is another problem building with SCons on Py3, which seems to be a
SCons bug. The workaround is to disable CacheDir in SConstruct. This has
been reported that one upstream to SCons.

Thanks to Hector Martin for the report and the patch.


git-svn-id: http://subversion.ffado.org/ffado@2775 2be59082-3212-0410-8809-b0798e1608f0